+++ apps/opentelemetry_api/src/otel_span.erl
@@ -41,7 +41,6 @@
 -define(is_recording(SpanCtx), SpanCtx =/= undefined andalso SpanCtx#span_ctx.is_recording =:= true).
 
 -type start_opts() :: #{attributes => opentelemetry:attributes(),
-                        sampler => term(),
                         links => opentelemetry:links(),
                         is_recording => boolean(),
                         start_time => opentelemetry:timestamp(),