+++ apps/opentelemetry_api/test/open_telemetry_test.exs
@@ -62,7 +62,6 @@ defmodule OpenTelemetryTest do
       kind: :internal,
       links: [],
       start_time: :opentelemetry.timestamp()
-      # sampler
     } do
       :ok
     end